# contracts/openapi.json
{
  "components": {
    "schemas": {
...
      "CurrentUserResponse": {
        "description": "The caller's own non-sensitive profile and effective authorisation.",
        "properties": {
          "auth_method": {
            "title": "Auth Method",
            "type": "string"
          },
          "display_name": {
            "title": "Display Name",
            "type": "string"
          },
          "email": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Email"
          },
          "facility_scope_ids": {
            "items": {
              "format": "uuid",
              "type": "string"
            },
            "title": "Facility Scope Ids",
            "type": "array"
          },
          "geography_scopes": {
            "items": {
              "$ref": "#/components/schemas/GeographyScopeSummary"
            },
            "title": "Geography Scopes",
            "type": "array"
          },
          "has_national_scope": {
            "title": "Has National Scope",
            "type": "boolean"
          },
          "is_synthetic": {
            "description": "True for development accounts. The interface must mark these visibly.",
            "title": "Is Synthetic",
            "type": "boolean"
          },
          "max_sensitivity": {
            "title": "Max Sensitivity",
            "type": "string"
          },
          "organisation_label": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Organisation Label"
          },
          "permissions": {
            "items": {
              "type": "string"
            },
            "title": "Permissions",
            "type": "array"
          },
          "roles": {
            "items": {
              "type": "string"
            },
            "title": "Roles",
            "type": "array"
          },
          "user_id": {
            "format": "uuid",
            "title": "User Id",
            "type": "string"
          },
          "username": {
            "title": "Username",
            "type": "string"
          }
        },
        "required": [
          "user_id",
          "username",
          "display_name",
          "email",
          "organisation_label",
          "roles",
          "permissions",
          "max_sensitivity",
          "geography_scopes",
          "facility_scope_ids",
          "has_national_scope",
          "auth_method",
          "is_synthetic"
        ],
        "title": "CurrentUserResponse",
        "type": "object"
      },
...
      "DevelopmentLoginRequest": {
        "description": "Request a synthetic development token. Non-production only.",
        "properties": {
          "username": {
            "maxLength": 128,
            "minLength": 1,
            "title": "Username",
            "type": "string"
          }
        },
        "required": [
          "username"
        ],
        "title": "DevelopmentLoginRequest",
        "type": "object"
      },
      "DevelopmentLoginResponse": {
        "properties": {
          "access_token": {
            "title": "Access Token",
            "type": "string"
          },
          "expires_at": {
            "format": "date-time",
            "title": "Expires At",
            "type": "string"
          },
          "token_type": {
            "default": "bearer",
            "title": "Token Type",
            "type": "string"
          },
          "warning": {
            "default": "Synthetic development session. Not valid in staging or production.",
            "title": "Warning",
            "type": "string"
          }
        },
        "required": [
          "access_token",
          "expires_at"
        ],
        "title": "DevelopmentLoginResponse",
        "type": "object"
      },
...
      "GeographyScopeSummary": {
        "properties": {
          "geography_unit_id": {
            "format": "uuid",
            "title": "Geography Unit Id",
            "type": "string"
          },
          "level": {
            "title": "Level",
            "type": "string"
          },
          "name": {
            "title": "Name",
            "type": "string"
          },
          "preferred_code": {
            "title": "Preferred Code",
            "type": "string"
          }
        },
        "required": [
          "geography_unit_id",
          "preferred_code",
          "level",
          "name"
        ],
        "title": "GeographyScopeSummary",
        "type": "object"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      },
...
    },
    "securitySchemes": {
      "MARS bearer token": {
        "scheme": "bearer",
        "type": "http"
      }
    }
  },
  "info": {
    "contact": {
      "name": "MARS implementation team"
    },
    "description": "MARS converts routine malaria data into explainable surveillance signals that a\nnamed person is accountable for investigating.\n\n**Scientific boundary.** Signals produced from routine e-register and HMIS data\nindicate patterns requiring investigation. They do not confirm antimalarial\nresistance. Routine data cannot distinguish recrudescence from reinfection,\nprove drug exposure or adherence, identify parasite genotype, or confirm\nmolecular markers. Externally confirmed findings - therapeutic efficacy studies\nand molecular results - are handled in a separate, separately governed lane. See\n`/api/v1/meta/evidence-lanes`.\n\n**Phase status.** This build covers phases 1 and 2: repository and runtime\nfoundation, database and governance foundation, authentication and\nauthorisation, and the organisation, facility and geography domain model. No\nencounter ingestion, episode construction, indicator, anomaly-detection, signal\nor investigation capability exists yet.\n",
    "title": "MARS - Malaria Antimalarial Resistance Surveillance",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
    "/api/v1/auth/dev/login": {
      "post": {
        "description": "Issue a short-lived synthetic token.\n\nThe account must already exist and be flagged synthetic. This route never\ncreates an account, so it cannot be used to mint a principal that the seeded\ndevelopment fixture did not define.",
        "operationId": "development_login_api_v1_auth_dev_login_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/DevelopmentLoginRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DevelopmentLoginResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Sign in as a synthetic development user",
        "tags": [
          "auth",
          "development"
        ]
      }
    },
...
    "/api/v1/auth/logout": {
      "post": {
        "description": "Record the logout.\n\nToken revocation is the identity provider's responsibility; MARS records the\nevent so the audit trail has both ends of the session.",
        "operationId": "logout_api_v1_auth_logout_post",
        "responses": {
          "204": {
            "description": "Successful Response"
          }
        },
        "security": [
          {
            "MARS bearer token": []
          }
        ],
        "summary": "End the current session",
        "tags": [
          "auth"
        ]
      }
    },
    "/api/v1/auth/me": {
      "get": {
        "description": "Return the caller's own non-sensitive profile.\n\nThe frontend uses this to decide what to render. It is a convenience, not a\ncontrol: every endpoint re-checks server-side, so a client that ignores this\nresponse gains nothing.",
        "operationId": "current_user_api_v1_auth_me_get",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CurrentUserResponse"
                }
              }
            },
            "description": "Successful Response"
          }
        },
        "security": [
          {
            "MARS bearer token": []
          }
        ],
        "summary": "The caller's profile and effective authorisation",
        "tags": [
          "auth"
        ]
      }
    },
...
  }
}